Communication from bracelet to app :)

+++ app/src/main/java/com/example/colormatchingbracelet/Bracelet/BraceletInformation.java
@@ -5,7 +5,7 @@ import android.graphics.Color;
 import java.util.List;
 
 public class BraceletInformation {
-    public float batteryPercentage;
+    public int batteryPercentage;
     public boolean ledStripPowerState;
     public int ledStripBrightness;
